native: LP64 tail — .cpp unsigned long->u32, JUTConsole -sizeof int cast
- JAIGFrameSequence/JKRExpHeap/MapDraw/MarioCap .cpp: unsigned long->u32 in the
  signatures whose headers already moved to u32 (vtable/decl match on LP64).
- JUTConsole.hpp: TLinkList<JUTConsole, -sizeof(JKRDisposer)> -- size_t is 64-bit
  on host so -sizeof overflowed the int offset param; cast -(int)sizeof.

(Build also passes -Wno-narrowing to match the Metrowerks toolchain, which did
not enforce C++11 narrowing-in-aggregate-init strictness; that is a compiler-flag
match, not a source change.)

+++ include/JSystem/JUtility/JUTConsole.hpp
@@ -151,7 +151,9 @@ public:
 	    ConsoleList;
 #else
 	// clangd does not support offsetof in template arguments.
-	typedef JGadget::TLinkList<JUTConsole, -sizeof(JKRDisposer)> ConsoleList;
+	// -(int) cast: TLinkList's offset param is int; on host size_t is 64-bit so
+	// -sizeof(...) is a huge u64 that overflows int (on GC size_t was 32-bit).
+	typedef JGadget::TLinkList<JUTConsole, -(int)sizeof(JKRDisposer)> ConsoleList;
 #endif
 
 private:
